requirement.txt
asgiref==3.8.1
Django==5.2.3
pillow==10.4.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
ttkbootstrap==1.13.11
tzdata==2025.2
